+++ eventmetrics/parsers/timeElapsedParser.go
@@ -26,11 +26,9 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/xmidt-org/interpreter/validation"
 
-	"github.com/go-kit/kit/log"
-	"github.com/go-kit/kit/log/level"
 	"github.com/xmidt-org/interpreter"
 	"github.com/xmidt-org/interpreter/history"
-	"github.com/xmidt-org/themis/xlog"
+	"go.uber.org/zap"
 )
 
 const (
@@ -45,6 +43,8 @@ const (
 	hardwareMetadataKey     = "/hw-model"
 	firmwareMetadataKey     = "/fw-name"
 	rebootReasonMetadataKey = "/hw-last-reboot-reason"
+
+	invalidIncomingMsg = "invalid incoming event"
 )
 
 // EventClient is an interface that provides a list of events related to a device.
@@ -77,13 +77,13 @@ type TimeElapsedParser struct {
 	incomingEvent EventInfo
 	searchedEvent EventInfo
 	name          string
-	logger        log.Logger
+	logger        *zap.Logger
 	client        EventClient
 	measures      Measures
 }
 
 // NewTimeElapsedParser creates a TimeElapsedParser from a TimeElapsedConfig or returns an error if there one cannot be created.
-func NewTimeElapsedParser(config TimeElapsedConfig, client EventClient, logger log.Logger, measures Measures, currentTime TimeFunc) (*TimeElapsedParser, error) {
+func NewTimeElapsedParser(config TimeElapsedConfig, client EventClient, logger *zap.Logger, measures Measures, currentTime TimeFunc) (*TimeElapsedParser, error) {
 	incomingEvent, err := NewEventInfo(config.IncomingEvent, currentTime)
 	if err != nil {
 		return nil, err
@@ -119,7 +119,7 @@ func NewTimeElapsedParser(config TimeElapsedConfig, client EventClient, logger l
 		finder:        finder,
 		incomingEvent: incomingEvent,
 		searchedEvent: searchedEvent,
-		logger:        logger,
+		logger:        logger.With(zap.String("parser", config.Name)),
 		client:        client,
 		measures:      measures,
 		name:          config.Name,
@@ -160,7 +160,7 @@ func (t *TimeElapsedParser) Parse(event interpreter.Event) {
 		}
 		histogram.With(labels).Observe(restartTime)
 	} else {
-		level.Error(t.logger).Log(xlog.ErrorKey(), "No histogram found for this time elapsed parser")
+		t.logger.Error("No histogram found for this time elapsed parser")
 	}
 }
 
@@ -188,23 +188,23 @@ func fixConfig(config TimeElapsedConfig, defaultValidFrom time.Duration) TimeEla
 func (t *TimeElapsedParser) calculateTimeElapsed(incomingEvent interpreter.Event) (float64, error) {
 	if valid, err := t.incomingEvent.Valid(incomingEvent); !valid {
 		if errors.Is(err, validation.ErrInvalidEventType) {
-			level.Info(t.logger).Log(xlog.MessageKey(), err, "event destination", incomingEvent.Destination)
+			t.logger.Debug(invalidIncomingMsg, zap.Error(err), zap.String("event destination", incomingEvent.Destination))
 		} else {
-			level.Error(t.logger).Log(xlog.ErrorKey(), err, "event destination", incomingEvent.Destination)
+			t.logger.Error(invalidIncomingMsg, zap.Error(err), zap.String("event destination", incomingEvent.Destination))
 		}
 		return -1, err
 	}
 
 	deviceID, err := incomingEvent.DeviceID()
 	if err != nil {
-		level.Error(t.logger).Log(xlog.ErrorKey(), err)
+		t.logger.Error(invalidIncomingMsg, zap.Error(err))
 		return -1, TimeElapsedCalculationErr{err: err, errLabel: idParseReason}
 	}
 
 	events := t.client.GetEvents(deviceID)
 	oldEvent, err := t.finder.Find(events, incomingEvent)
 	if err != nil {
-		t.logErrWithEventDetails(err, incomingEvent)
+		t.logErrWithEventDetails(invalidIncomingMsg, err, incomingEvent)
 		return -1, err
 	}
 
@@ -217,7 +217,7 @@ func (t *TimeElapsedParser) calculateTimeElapsed(incomingEvent interpreter.Event
 
 	if timeElapsed <= 0 {
 		err = TimeElapsedCalculationErr{timeElapsed: timeElapsed, oldEvent: oldEvent, errLabel: invalidTimeCalculatedReason}
-		t.logErrWithEventDetails(err, incomingEvent)
+		t.logErrWithEventDetails("invalid calculation", err, incomingEvent)
 		return -1, err
 	}
 
@@ -225,15 +225,15 @@ func (t *TimeElapsedParser) calculateTimeElapsed(incomingEvent interpreter.Event
 }
 
 // logs errors with events' transaction uuids.
-func (t *TimeElapsedParser) logErrWithEventDetails(err error, incomingEvent interpreter.Event) {
+func (t *TimeElapsedParser) logErrWithEventDetails(msg string, err error, incomingEvent interpreter.Event) {
 	deviceID, _ := incomingEvent.DeviceID()
 	var eventErr ErrorWithEvent
 	if errors.As(err, &eventErr) {
 		if len(eventErr.Event().TransactionUUID) > 0 {
-			level.Error(t.logger).Log(xlog.ErrorKey(), err, "deviceID", deviceID, "incoming event", incomingEvent.TransactionUUID, "old event", eventErr.Event().TransactionUUID)
+			t.logger.Error(msg, zap.Error(err), zap.String("deviceID", deviceID), zap.String("incoming event", incomingEvent.TransactionUUID), zap.String("comparison event", eventErr.Event().TransactionUUID))
 			return
 		}
 	}
 
-	level.Error(t.logger).Log(xlog.ErrorKey(), err, "deviceID", deviceID, "incoming event", incomingEvent.TransactionUUID)
+	t.logger.Error(msg, zap.Error(err), zap.String("deviceID", deviceID), zap.String("incoming event", incomingEvent.TransactionUUID))
 }