+++ eventmetrics/parsers/timeElapsedParser_test.go
@@ -8,8 +8,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/go-kit/kit/log"
-	logging "github.com/go-kit/kit/log"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/testutil"
 	"github.com/stretchr/testify/assert"
@@ -18,6 +16,8 @@ import (
 	"github.com/xmidt-org/interpreter/history"
 	"github.com/xmidt-org/interpreter/validation"
 	"github.com/xmidt-org/touchstone/touchtest"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
 )
 
 type testFinder struct {
@@ -31,7 +31,7 @@ func TestNewTimeElapsedParser(t *testing.T) {
 	now, err := time.Parse(time.RFC3339Nano, "2021-03-02T18:00:01Z")
 	assert.Nil(t, err)
 	currTime := func() time.Time { return now }
-	logger := log.NewNopLogger()
+	logger := zap.NewNop()
 	mockClient := new(mockEventClient)
 
 	tests := []struct {
@@ -354,7 +354,7 @@ func testParseCalculateErr(t *testing.T) {
 			}
 			parser := TimeElapsedParser{
 				incomingEvent: EventInfo{Validator: mockVal},
-				logger:        logging.NewNopLogger(),
+				logger:        zap.NewNop(),
 				measures:      m,
 				name:          "TEP_test",
 			}
@@ -424,7 +424,7 @@ func testParseNoHWFWErr(t *testing.T) {
 			parser := TimeElapsedParser{
 				searchedEvent: EventInfo{CalculateUsing: Birthdate, Validator: mockVal},
 				incomingEvent: EventInfo{CalculateUsing: Birthdate, Validator: mockVal},
-				logger:        logging.NewNopLogger(),
+				logger:        zap.NewNop(),
 				client:        mockClient,
 				finder:        mockFinder,
 				measures:      m,
@@ -516,7 +516,7 @@ func testParseSuccess(t *testing.T) {
 			parser := TimeElapsedParser{
 				searchedEvent: EventInfo{CalculateUsing: Birthdate, Validator: mockVal},
 				incomingEvent: EventInfo{CalculateUsing: Birthdate, Validator: mockVal},
-				logger:        logging.NewNopLogger(),
+				logger:        zap.NewNop(),
 				client:        mockClient,
 				finder:        mockFinder,
 				measures:      m,
@@ -574,15 +574,20 @@ func TestLogErrWithEventDetails(t *testing.T) {
 		t.Run(tc.description, func(t *testing.T) {
 			assert := assert.New(t)
 			buf := &bytes.Buffer{}
-			logger := log.NewJSONLogger(buf)
+			ws := zapcore.AddSync(buf)
+			core := zapcore.NewCore(zapcore.NewJSONEncoder(zap.NewDevelopmentEncoderConfig()), ws, zapcore.DebugLevel)
+			logger := zap.New(core)
+			msg := "testMsg"
+
 			parser := TimeElapsedParser{
 				logger: logger,
 			}
-			parser.logErrWithEventDetails(tc.err, tc.event)
+			parser.logErrWithEventDetails(msg, tc.err, tc.event)
 			logStr := buf.String()
 			for _, logVal := range tc.expectedJSONLog {
 				assert.Contains(logStr, logVal)
 			}
+			assert.Contains(logStr, msg)
 		})
 	}
 }
@@ -590,7 +595,7 @@ func TestLogErrWithEventDetails(t *testing.T) {
 func testInvalidIncomingEvent(t *testing.T) {
 	assert := assert.New(t)
 	mockVal := new(mockValidator)
-	logger := log.NewNopLogger()
+	logger := zap.NewNop()
 	invalidEventErr := errors.New("invalid event")
 	incomingEvent := interpreter.Event{
 		Destination: "some-destination",
@@ -613,7 +618,7 @@ func testInvalidIncomingEvent(t *testing.T) {
 func testDeviceIDErr(t *testing.T) {
 	assert := assert.New(t)
 	mockVal := new(mockValidator)
-	logger := log.NewNopLogger()
+	logger := zap.NewNop()
 	incomingEvent := interpreter.Event{
 		Destination: "some-destination",
 	}
@@ -630,7 +635,7 @@ func testDeviceIDErr(t *testing.T) {
 
 func testFinderErr(t *testing.T) {
 	assert := assert.New(t)
-	logger := log.NewNopLogger()
+	logger := zap.NewNop()
 	testErr := errors.New("test error")
 	incomingEvent := interpreter.Event{
 		Destination: "event:device-status/mac:112233445566/fully-manageable",
@@ -658,7 +663,7 @@ func testFinderErr(t *testing.T) {
 func testCalculations(t *testing.T) {
 	now, err := time.Parse(time.RFC3339Nano, "2021-03-02T18:00:01Z")
 	assert.Nil(t, err)
-	logger := log.NewNopLogger()
+	logger := zap.NewNop()
 	mockVal := new(mockValidator)
 	mockClient := new(mockEventClient)
 	mockFinder := new(mockFinder)